Avoid executing function-sidecar executable in integration test
Fixes https://github.com/projectriff/function-sidecar/issues/45.
