+++ README.adoc
@@ -11,5 +11,5 @@ make dockerize
 == Tests (including integration with Kafka)
 Assuming kafka is running locally on port 9092,
 ```
-KAFKA_BROKER=localhost:9092 make clean test
+KAFKA_BROKER=localhost:9092 make test
 ```
